# Optimized the binary encoder for a 2x speedup