… upb/pb/encoder.c