This PR optimizes the varint-encoding path, in particular the fast path for one-byte varints.

There is some modest code size growth, but a roughly 2x speed improvement:

```
name                       old time/op  new time/op  delta        
ArenaOneAlloc              21.1ns ± 0%  20.5ns ± 1%   -3.06%  (p=0.000 n=12+12)
ArenaInitialBlockOneAlloc  5.25ns ± 0%  5.25ns ± 0%     ~     (p=0.833 n=11+12)
ParseDescriptorNoHeap      16.4µs ± 0%  16.9µs ± 0%   +2.48%  (p=0.000 n=12+12)
ParseDescriptor            17.0µs ± 1%  17.3µs ± 1%   +1.69%  (p=0.000 n=12+12)
SerializeDescriptor        21.6µs ± 0%  10.7µs ± 0%  -50.66%  (p=0.000 n=12+12)


    FILE SIZE        VM SIZE   
 --------------  --------------
   +16%    +815    +16%    +749    upb/encode.c
      [NEW] +3.00Ki   [NEW] +2.96Ki    encode_message
      [NEW] +1.12Ki   [NEW] +1.08Ki    encode_scalar
      [NEW]    +322   [NEW]    +280    encode_growbuffer
      [NEW]    +293   [NEW]    +251    encode_fixedarray
      [NEW]    +214   [NEW]    +170    encode_bytes.part.0
      [NEW]    +205   [NEW]    +163    encode_longvarint
      [NEW]    +131   [NEW]     +91    encode_varint64
      [NEW]    +120   [NEW]     +85    encode_tag
       +38%     +74    +46%     +74    upb_encode
      [NEW]     +32   [NEW]      +1    ch.5007
      [DEL]     -32   [DEL]      -1    ch.4956
      [DEL]    -195   [DEL]    -157    upb_put_bytes
      [DEL]    -274   [DEL]    -235    upb_put_varint
      [DEL]    -275   [DEL]    -232    upb_put_fixedarray
      [DEL]    -289   [DEL]    -243    upb_encode_growbuffer
      [DEL]    -985   [DEL]    -938    upb_encode_scalarfield
      [DEL] -2.68Ki   [DEL] -2.64Ki    upb_encode_message
  +1.7%     +35   +1.7%     +35    [section .text]
 -19.8%    -786   [ = ]       0    [Unmapped]
  +0.0%     +64   +0.6%    +784    TOTAL
```

This PR also includes a benchmark for encoding, as well as updates to the benchmarking script that produces the above output.

I also renamed the functions in the encoder to follow a more regular `encode_*()` scheme, and renamed some symbols in the old encoder to avoid collisions. The old encoder is on its last legs, and will be deleted once Ruby is migrated to `upb_msg`.